+++ CHANGES.md
@@ -1,6 +1,12 @@
 IRC Client Daemon CHANGES
 =========================
 
+irccd 4.1.0 ????-??-??
+======================
+
+- Add `irc_bot_pollable_add` to insert a custom interface into the main irccd
+  loop.
+
 irccd 4.0.1 2022-02-07
 ======================